acabada la camara
la pinche camara puta conchuida estercolera hija de su putamadre zorra

+++ TPV2/src/game/Character.cpp
@@ -393,13 +393,13 @@ void Character::draw(SDL_Rect* camera)
 	SDL_Rect aux = hurtbox;
 
 	aux.x -= camera->x;
-	aux.x *= (manager->GetActualWidth() / camera->w);
+	aux.x *= (manager->GetActualWidth() / (float)camera->w);
 
 	aux.y -= camera->y;
-	aux.y *= (manager->GetActualHeight() / camera->h);
+	aux.y *= (manager->GetActualHeight() / (float)camera->h);
 
-	aux.w *= (manager->GetActualWidth() / camera->w);
-	aux.h *= (manager->GetActualHeight() / camera->h);
+	aux.w *= (manager->GetActualWidth() / (float)camera->w);
+	aux.h *= (manager->GetActualHeight() / (float)camera->h);
 
 	SDL_RenderDrawRect(sdl->renderer(), &aux);
 }
