+++ TPV2/src/game/Utils/Particle.cpp
@@ -156,13 +156,13 @@ void Particle::draw(SDL_Rect* camera)
 	SDL_Rect aux = dest;
 
 	aux.x -= camera->x;
-	aux.x *= (manager->GetActualWidth() / camera->w);
+	aux.x *= (manager->GetActualWidth() / (float)camera->w);
 
 	aux.y -= camera->y;
-	aux.y *= (manager->GetActualHeight() / camera->h);
+	aux.y *= (manager->GetActualHeight() / (float)camera->h);
 
-	aux.w *= (manager->GetActualWidth() / camera->w);
-	aux.h *= (manager->GetActualHeight() / camera->h);
+	aux.w *= (manager->GetActualWidth() / (float)camera->w);
+	aux.h *= (manager->GetActualHeight() / (float)camera->h);
 
 	if (dir < 0)
 	{
